// web/src/api/generated/models/banner-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 轮播图响应
 */
export interface BannerResponse {
    /**
     * 轮播图ID
     */
    'id'?: number;
    /**
     * 标题
     */
    'title'?: string;
    /**
     * 图片URL
     */
    'imageUrl'?: string;
    /**
     * 跳转类型: 0-无跳转, 1-内部路由, 2-外部链接
     */
    'linkType'?: number;
    /**
     * 跳转类型描述
     */
    'linkTypeDesc'?: string;
    /**
     * 跳转URL/路由
     */
    'linkUrl'?: string;
    /**
     * 排序权重
     */
    'sort'?: number;
    /**
     * 状态: 0-草稿, 1-已发布, 2-已下线
     */
    'status'?: number;
    /**
     * 状态描述
     */
    'statusDesc'?: string;
    /**
     * 开始展示时间
     */
    'startTime'?: string;
    /**
     * 结束展示时间
     */
    'endTime'?: string;
    /**
     * 创建者ID
     */
    'adminId'?: number;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
}