// web/src/api/generated/models/update-banner-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 更新轮播图请求
 */
export interface UpdateBannerRequest {
    /**
     * 轮播图ID
     */
    'id': number;
    /**
     * 标题
     */
    'title': string;
    /**
     * 图片URL
     */
    'imageUrl': string;
    /**
     * 跳转类型: 0-无跳转, 1-内部路由, 2-外部链接
     */
    'linkType'?: number;
    /**
     * 跳转URL/路由
     */
    'linkUrl'?: string;
    /**
     * 排序权重，值越大越靠前
     */
    'sort'?: number;
    /**
     * 开始展示时间
     */
    'startTime'?: string;
    /**
     * 结束展示时间
     */
    'endTime'?: string;
    /**
     * 状态: 0-草稿, 1-已发布, 2-已下线
     */
    'status'?: number;
}