// web/src/api/generated/models/query-banner-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 查询轮播图请求
 */
export interface QueryBannerRequest {
    /**
     * 标题关键词
     */
    'title'?: string;
    /**
     * 状态: 0-草稿, 1-已发布, 2-已下线
     */
    'status'?: number;
    /**
     * 创建者ID
     */
    'adminId'?: number;
    /**
     * 页码
     */
    'pageNum'?: number;
    /**
     * 每页数量
     */
    'pageSize'?: number;
}